Split out on_create.rs
